refactor: relocate services and enhance client filters
