+++ src/app/(auth)/_components/login-form/index.tsx
@@ -12,7 +12,7 @@ import { toast } from "sonner";
 import { useRouter } from "next/navigation";
 
 // Domain Entities/DTOs/Interfaces
-import { useAuth } from "@/app/services/auth";
+import { useAuth } from "@/services/auth";
 
 // Utils/Helpers
 import {
