+++ src/app/admin/_components/admin-header.tsx
@@ -5,7 +5,7 @@ import Link from "next/link";
 import { useRouter } from "next/navigation";
 
 // Services
-import { useAuth } from "@/app/services/auth";
+import { useAuth } from "@/services/auth";
 
 // Components
 import {
